% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_bigdata.R
\name{genee_bigdata}
\alias{genee_bigdata}
\title{genee: An Association Test for Aggregated Sets of SNP-Level Summary Statistics.}
\usage{
genee_bigdata(mydata, ld_matrix, ld_path, alpha, upper, lower,
  prior_weight, gene_list, nfolds, lambda.min, nlambda)
}
\arguments{
\item{mydata}{A matrix of four columns. The first column specifies the chromosome for each SNP. The second column specifies the name for each SNP. The third column specifies the physical position for each SNP. The fourth column specifies the GWAS effect size estimator for each SNP.}

\item{ld_path}{Path to the ld matrix memory mapping file (.desc, .bin) including the prefix of the file name. Make sure that .desc and .bin are in the same directory.}

\item{alpha}{A tuning parameter determining the type of regularized regression. Default is alpha = 0.5 which uses the Elastic Net solution. More specifically, alpha = 0: Ridge Regression; 0 < alpha < 1: Elastic Net; alpha = 1: LASSO. Set alpha = -1 to use the original summary statistics with no regularization or shrinkage.}

\item{upper}{The number of base pairs downstream of the gene to be included. Default is 50000.}

\item{lower}{The number of base pairs upstream of the gene to be included. Default is 50000.}

\item{prior_weight}{A vector specifying the prior weight for each SNP. Default assumes equal weights by defining prior_weight to be a vector containing 1's.}

\item{gene_list}{A list where each element represents a vector containing the indices of all SNPs in one set (e.g., a set could refer to a gene). If it's not provided by user, this will be derived using the hg19 gene list from UCSC browser.}

\item{nfolds}{A number indicates how many folds to do for cross validation for regularize regression. The default is 10. If it's fed with 0, then it will perform no cross validation and choose lambda with smallest penalty.}

\item{lambda.min}{The smallest lambda to use for regularize regression, as a fraction of the largest lambda. Default is 0.0001.}

\item{nlambda}{The number of lambda values to use. Default is 100.}

\item{ld}{A matrix of pairwise correlations statistics (i.e., r).}
}
\value{
If a gene list is not provided, genee will return a matrix containing the following 7 columns: 1. chromosome number (chr); 2. gene name; 3. start position; 4. end position; 5. number of SNPs in the gene/set (nsnp); 6. gene/set test statistics (test_q); and 7. gene/set p-value (pval). If a gene list is provided, genee will return a matrix with two columns: 1. gene/set test statistics (test_q); and 2. gene/set p-value (pval).
}
\description{
genee: An Association Test for Aggregated Sets of SNP-Level Summary Statistics.
}
\examples{
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
write.table(ld, file = "myld", row.names = FALSE, col.names = FALSE, quote = FALSE, sep = " ")
ld_name = "myld"
prepare_LD(ld_name)
ld_path = "myld"
prior_weight = c(1, 1, 1)
gene_list=list()
gene_list[[1]]=c(1, 2)
gene_list[[2]]=c(2, 3)
chr = c(1, 1, 1)
snpnames = c('rs1', 'rs2', 'rs3')
pos = c(100, 200, 300)
betas = c(0.1, 0.2, 0.1)
mydata = matrix(nrow = 3, ncol = 4)
mydata[,1] = chr
mydata[,2] = snpnames
mydata[,3] = pos
mydata[,4] = betas
genee(mydata, ld, ld_path, alpha =0.5, gene_list = gene_list, 3, 0.001, 50)
}
