% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_EM.R
\name{genee_EM}
\alias{genee_EM}
\title{EM algorithm to find mixture components and epsilon-genic effects.}
\usage{
genee_EM(betas)
}
\arguments{
\item{betas}{A vector of effect size estimators (OLS betas or regularized betas).}
}
\value{
The threshold for epsilon-genic effects (i.e., the second largest variance of all components).
}
\description{
EM algorithm to find mixture components and epsilon-genic effects.
}
\examples{
betas = c(rnorm(1000, 1, 1), rnorm(2000, 2,0))
genee_EM(betas)
}
