% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_loop.R
\name{genee_loop}
\alias{genee_loop}
\title{Loop through all gene sets and derive test statistics and p-values for each set.}
\usage{
genee_loop(betas, ld, epsilon_effect, prior_weight, gene_list)
}
\arguments{
\item{betas}{A vector of effect size estimators (OLS betas or regularized betas).}

\item{ld}{A matrix of pairwise correlations statistics (i.e., r).}

\item{epsilon_effect}{The threshold for epsilon-genic effects (i.e., the second largest variance component).}

\item{prior_weight}{A vector specifying the prior weight for each SNP.}

\item{gene_list}{A list where each element represents a vector containing the indices of all SNPs in one set (e.g., a set could refer to a gene).}
}
\value{
A list where the first element is a vector of test statistics, and the second element is a vector of p values.
}
\description{
Loop through all gene sets and derive test statistics and p-values for each set.
}
\examples{
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
betas = c(0.1, 0.2, 0.1)
prior_weight = c(1, 1, 1)
gene_list=list()
gene_list[[1]]=c(1, 2)
gene_list[[2]]=c(2, 3)
genee_loop(betas, ld, 1e-4, prior_weight, gene_list)
}
