% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_regularize_bigdata.R
\name{genee_regularize_bigdata}
\alias{genee_regularize_bigdata}
\title{genee test using regularized effect size estimators.}
\usage{
genee_regularize_bigdata(betas_ols, alpha, ld, ld_path, prior_weight,
  gene_list, nfolds, lambda.min, nlambda)
}
\arguments{
\item{betas_ols}{A vector of effect size estimators (OLS betas).}

\item{alpha}{A tuning parameter determining the type of regularization. More specifically, alpha = 0: Ridge Regression; 0 < alpha < 1: Elastic Net; alpha = 1: LASSO.}

\item{ld}{A matrix of pairwise correlations statistics (i.e., r).}

\item{ld_path}{Path to the ld matrix memory mapping file (.desc, .bin) including the prefix of the file name. Make sure that .desc and .bin are in the same directory.}

\item{prior_weight}{A vector specifying the prior weight for each SNP.}

\item{gene_list}{A list where each element represents a vector containing the indices of all the SNPs in one set (e.g. a set could refer to a gene).}

\item{nfolds}{A number indicates how many folds to do for cross validation for regularize regression. The default is 10. If it's fed with 0, then it will perform no cross validation and choose lambda with smallest penalty.}

\item{lambda.min}{The smallest lambda to use for regularize regression, as a fraction of the largest lambda. Default is 0.0001.}

\item{nlambda}{The number of lambda values to use. Default is 100.}
}
\value{
A list where the first element is a vector of test statistics and the second element is a vector of p values.
}
\description{
genee test using regularized effect size estimators.
}
\examples{
x1 = c(0, 1, 0)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
write.table(ld, file = "myld", row.names = FALSE, col.names = FALSE, quote = FALSE, sep = " ")
ld_name = "myld"
prepare_LD(ld_name)
ld_path = "myld"
betas = c(2, 3, -2)
prior_weight = c(1, 1, 1)
gene_list=list()
gene_list[[1]]=c(1, 2)
gene_list[[2]]=c(2, 3)
genee_regularize_bigdata(betas, 0.5, ld_path, prior_weight, gene_list, 5, 0.001, 50)
}
