% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_regularize_regression_bigdata.R
\name{genee_regularize_regression_bigdata}
\alias{genee_regularize_regression_bigdata}
\title{Regularized Regression.}
\usage{
genee_regularize_regression_bigdata(betas, ld_path, alpha, nfolds,
  lambda.min, nlambda)
}
\arguments{
\item{betas}{A vector of GWAS effect size estimators (OLS betas).}

\item{ld_path}{Path to the ld matrix memory mapping file (.desc, .bin) including the prefix of the file name. Make sure that .desc and .bin are in the same directory.}

\item{alpha}{A tuning parameter determining the type of regularization. More specifically, alpha = 0: Ridge Regression; 0 < alpha < 1: Elastic Net; alpha = 1: LASSO.}

\item{nfolds}{A number indicates how many folds to do for cross validation for regularize regression. The default is 10. If it's fed with 0, then it will perform no cross validation and choose lambda with smallest penalty.}

\item{lambda.min}{The smallest lambda to use for regularize regression, as a fraction of the largest lambda. Default is 0.0001.}

\item{nlambda}{The number of lambda values to use. Default is 100.}
}
\value{
A vector of regularized effect size estimators.
}
\description{
Regularized Regression.
}
\examples{
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
write.table(ld, file = "myld", row.names = FALSE, col.names = FALSE, quote = FALSE, sep = " ")
ld_name = "myld"
prepare_LD(ld_name)
ld_path = "myld"
betas = c(0.1, 0.2, 0.1)
genee_regularize_regression_bigdata(betas, ld_path, 0.5, 5, 0.001, 50)
}
