% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_test.R
\name{genee_test}
\alias{genee_test}
\title{Hypothesis Testing for One Set of SNPs}
\usage{
genee_test(gene, ld, betas, epsilon_effect, prior_weight)
}
\arguments{
\item{gene}{A vector containing the indices of all the SNPs in the set (e.g., a set may refer to a gene).}

\item{ld}{A matrix of pairwise correlations statistics (i.e., r).}

\item{betas}{A vector of effect size estimators (OLS betas or regularized betas).}

\item{epsilon_effect}{The threshold for epsilon-genic effects (i.e., the second largest varaince of all components).}

\item{prior_weight}{A vector specifying the prior weight for each SNP.}
}
\value{
A vector containing test statistic and p-value for the set.
}
\description{
Hypothesis Testing for One Set of SNPs
}
\examples{
gene = c(1,2)
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
betas = c(0.1, 0.2, 0.1)
prior_weight = c(1, 1, 1)
genee_test(gene, ld, betas, 1e-4, prior_weight)
}
