% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ld.R
\name{prepare_LD}
\alias{prepare_LD}
\title{Preparing LD memory mapping file.}
\usage{
prepare_LD(ld_name)
}
\arguments{
\item{ld_name}{The name of LD file.}
}
\value{
Write out the ld memory mapped file (.desc, .bin).
}
\description{
Preparing LD memory mapping file.
}
\examples{
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
write.table(ld, file = "myld", row.names = FALSE, col.names = FALSE, quote = FALSE, sep = " ")
ld_name = "myld"
prepare_LD(ld_name)
}
