% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genee_regularize_regression.R
\name{genee_regularize_regression}
\alias{genee_regularize_regression}
\title{Regularized Regression.}
\usage{
genee_regularize_regression(betas, ld, alpha)
}
\arguments{
\item{betas}{A vector of GWAS effect size estimators (OLS betas).}

\item{ld}{A matrix of pairwise correlations statistics (i.e., r).}

\item{alpha}{A tuning parameter determining the type of regularization. More specifically, alpha = 0: Ridge Regression; 0 < alpha < 1: Elastic Net; alpha = 1: LASSO.}
}
\value{
A vector of regularized effect size estimators.
}
\description{
Regularized Regression.
}
\examples{
x1 = c(0, 1, 1)
x2 = c(0, 1, 2)
x3 = c(1, 1, 2)
x = cbind(x1 ,x2, x3)
ld = cor(x)
betas = c(0.1, 0.2, 0.1)
genee_regularize_regression(betas, ld, 0.5)
}
